Treat mutants surviving as an error
Surviving mutants would typically indicate insufficient test coverage.
As such, it makes sense for Mull to treat it as an error and exit with a
non-zero status.

This changes the default behavior of mull-runner which up until this
point has exited with zero regardless of how many mutants survived. This
behavior can be "restored" by passing the --allow-surviving CLI argument
to mull-runner.

Since the mutation scores for both fmtlib and openssl are very low, the
--allow-surviving flag is passed when running the integration tests

This does not treat the absence of mutants (i.e. infinite mutation
score) as an error